requirements.txt
ipfshttpclient==0.8.0
RPi.GPIO==0.7.0